backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
python-dateutil>=2.8
django-cors-headers>=3.13
